requirements.txt
django==1.5
mysql_python
django-crispy-forms
django-xadmin
requests
apscheduler
tablib
xlwt
xlrd